Support floating-point tokens. (#3)
* Support floating-point tokens.

This commit extends the support for reading numeric values by permitting
floating-point tokens of the forms listed below:

```
    1.2
    1.
    .2
    1e3
    1e+3
    1e-3
    1.2e3
    1.2e-3
    1.2e+3
    .2e3
    .2e-3
    .2e+3
```

This is done by replacing `read_integer()` with `read_number`, returning
tokens of kind `Integer`, `FloatingPoint`, or `Unknown`.

The `Unknown` token kind is returned when attempting to tokenize a
floating-point value fails due to unexpected alphanumeric chars. In this
case, the general "junk" tokenizer cannot be used because we may have
accepted one of `.+-eE` in the token already. One failure mode that was
already seen (and for which a test exists) is the token `2.1e`. This
should yield an `Unknown` token to skip past the junk, but with the
general junk parser it would return an Integer token `2`, a Period
token, then a `1e` Unknown token.

Arguably it is a design decision to capture this as a single junk token
rather than two valid tokens and one invalid. It seems more likely that
this approach will lead to better diagnostics later in translation,
compared to accepting a valid integer and period token and not reporting
them in the error message that would be generated for the `1e` junk
token. It also avoids the possible interpretation of `2.1e3.2` as
something like
```
Integer(2)
Period(.)
FloatingPoint(1e3)
FloatingPoint(.2)
```
instead preferring
```
Unknown(2.1e3)
FloatingPoint(.2)
```
which is still incorrect, but generates an `Unknown` from which an error
could be triggered, rather than making valid tokens out of what is
obviously an erroneous sequence of characters.

* Address review comments.

- Use grouping in the matchers.
- Remove excess blank lines.
- Change the way we process junk tokens in the f.p. tokeniser to read
  all chars that could be considered part of the representation of a
  number, thus ensuring junk tokens are visually nonsensical, not just
  contextually nonsensical.
- Inline `p` on line 164 (no further references to this var in the
  function).
- Add tests for more complex valid and invalid floating point sequences.