Handle unary operators `+` and `-` for integer and floating-point expressions.
This changeset provides support for expressions of the form `+1`, `-1`,
`+3.14159` and `-3.14159` when parsing integer or floating-point
expressions.

It adds two new variants, `UnaryPlusExpression` and
`UnaryMinusExpression` which can each hold either an `IntegerExpression`
or a `FloatingPointExpression`, and extends `ExpressionStatement` to
support these two new variants.

This completes the support for representing signed numbers, since the
tokenizer parses the preceding `+` or `-` operator separately to the
numeric value.