Factor complex logic out of main token matching. (#77)
This commit factors out two more complex if/else blocks from the matcher
for tokens. This has two effects:

1. The main matching logic becomes easier to read without nested if/else
   blocks for more complex matching.

2. `cargo tarpaulin` no longer reports lines are not covered because
   there is no executable code on that line. By moving the complex logic
   into a function, and putting a call to the function on the same line
   as the pattern, `tarpaulin` now correctly sees that it is covered.

Note that the equivalent refactoring was not done for the nested match
of `!=`, `>=` and `<=` as these are simple matches with no `if`
conditions, and there is benefit to having the full matching logic for
all such operators/symbols expressed directly in the matcher block.

There is a minor hit to locality of knowledge; that is, finding the
logic for tokenizing certain things requires more searching, but the
function names should be sufficiently descriptive to make this
near-trivial. For example, instead of the inlined logic to distinguish a
period token from a `.123` style decimal literal, the function is named
`read_period_or_decimal_token`. This should still provide sufficient
context for readers of the main match block to find the relevant code
further down.